+++ src/livePreviewPostProcessor.ts
@@ -3,68 +3,7 @@ import { Compartment, Extension } from "@codemirror/state";
 import { Decoration, DecorationSet, EditorView, PluginValue, ViewPlugin, ViewUpdate, WidgetType } from "@codemirror/view";
 import { editorLivePreviewField } from "obsidian";
 import VariablesPlugin from "./main";
-
-class VarWidget extends WidgetType {
-    constructor(readonly value: string) { super() }
-
-    eq(other: VarWidget) { return other.value == this.value }
-
-    toDOM() {
-        let wrap = document.createElement("span");
-        wrap.innerHTML = this.value;
-        return wrap
-    }
-
-    ignoreEvent() { return false }
-}
-
-function getLivePreviewRanges(view: EditorView) {
-
-    let selectedLines = new Set();
-    for (let { from, to } of view.state.selection.ranges) {
-        //algo can likely be improved for large selections
-        for (let pos = from; pos <= to; pos++) {
-            let line = view.state.doc.lineAt(pos);
-            selectedLines.add(line.number);
-        }
-    }
-
-    let livePreviewLines = new Set<number>();
-    for (let { from, to } of view.visibleRanges) {
-        for (let pos = from; pos <= to; pos++) {
-
-            let line = view.state.doc.lineAt(pos);
-            if (!selectedLines.has(line.number)) {
-                livePreviewLines.add(line.number);
-            }
-        }
-    }
-
-    let livePreviewRanges = [];
-    for (let lineNumber of livePreviewLines) {
-        let line = view.state.doc.line(lineNumber)
-        livePreviewRanges.push({ from: line.from, to: line.to });
-    }
-
-    return livePreviewRanges;
-}
-
-const debugTheme = EditorView.baseTheme({
-    "&light .cm-debug": { backgroundColor: "#597f7f" },
-    "&dark .cm-debug": { backgroundColor: "#597f7f" }
-})
-const debugStripe = Decoration.line({
-    attributes: { class: "cm-debug" }
-})
-
-export function livePreviewPostProcessorPlugin(plugin: VariablesPlugin): Extension {
-    return [
-        debugTheme,
-        ViewPlugin.define(v => new LivePreviewPostProcessor(v, plugin), {
-            decorations: v => v.decorations
-        })
-    ]
-}
+import { getLivePreviewRanges } from "./utils";
 
 class LivePreviewPostProcessor implements PluginValue {
     obsPlugin: VariablesPlugin;
@@ -80,7 +19,7 @@ class LivePreviewPostProcessor implements PluginValue {
         let builder = new RangeSetBuilder<Decoration>();
         if (update.state.field(editorLivePreviewField)) {
             for (let { from, to } of getLivePreviewRanges(update.view)) {
-                if (this.obsPlugin.debugMode) {
+                if (this.obsPlugin.settings.debugMode) {
                     let fromLine = update.view.state.doc.lineAt(from)
                     builder.add(fromLine.from, fromLine.from, debugStripe);
                 }
@@ -121,4 +60,35 @@ class LivePreviewPostProcessor implements PluginValue {
         }
         this.decorations = builder.finish();
     }
+}
+
+class VarWidget extends WidgetType {
+    constructor(readonly value: string) { super() }
+
+    eq(other: VarWidget) { return other.value == this.value }
+
+    toDOM() {
+        let wrap = document.createElement("span");
+        wrap.innerHTML = this.value;
+        return wrap
+    }
+
+    ignoreEvent() { return false }
+}
+
+const debugTheme = EditorView.baseTheme({
+    "&light .cm-debug": { backgroundColor: "#597f7f" },
+    "&dark .cm-debug": { backgroundColor: "#597f7f" }
+});
+const debugStripe = Decoration.line({
+    attributes: { class: "cm-debug" }
+});
+
+export function livePreviewPostProcessorPlugin(plugin: VariablesPlugin): Extension {
+    return [
+        debugTheme,
+        ViewPlugin.define(v => new LivePreviewPostProcessor(v, plugin), {
+            decorations: v => v.decorations
+        })
+    ]
 }
